fixed problems with rpm packaging

+++ pkg/deb/lib/postrm.sh
@@ -1,24 +1,30 @@
 #!/bin/bash
 
+VERSION="__VERSION__"
+
+if [ ! -f /usr/local/lib/kaveh-cluster-ctrl/VERSION ]; then
+  echo "${VERSION}" > /usr/local/lib/kaveh-cluster-ctrl/VERSION
+fi
+
+CUR_VERSION="$(cat /usr/local/lib/kaveh-cluster-ctrl/VERSION)"
 HAS="$(/bin/systemctl list-unit-files | grep kaveh-cluster-ctrl.service | awk '{print $1}')"
 
-if [ "${HAS}" == "kaveh-cluster-ctrl.service" ]; then
+if [ "${VERSION}" == "${CUR_VERSION}" ]; then
+  if [ "${HAS}" == "kaveh-cluster-ctrl.service" ]; then
     /bin/systemctl stop kaveh-cluster-ctrl.service || true
     /bin/systemctl disable kaveh-cluster-ctrl.service || true
+  fi
+
+  /bin/rm -f /lib/systemd/system/kaveh-cluster-ctrl.service
+
+  /bin/systemctl daemon-reload
+
+  /bin/rm -f /usr/local/bin/kluster-cli || true
+  /bin/rm -f /usr/local/bin/kluster-agent || true
+  /bin/rm -f /usr/local/lib/kaveh-cluster-ctrl/node/bin/npm
+  /bin/rm -f /usr/local/lib/kaveh-cluster-ctrl/node/bin/npx
+  /bin/rm -f /usr/local/lib/kaveh-cluster-ctrl/VERSION
+  /bin/rm -f /etc/kaveh-cluster-ctrl.conf
+  /bin/rm -rf /var/lib/kaveh-cluster-ctrl
 fi
 
-/bin/rm -f /lib/systemd/system/kaveh-cluster-ctrl.service
-
-/bin/systemctl daemon-reload
-
-HAS2="$(/bin/systemctl list-unit-files | grep kaveh-cluster-ctrl.service | awk '{print $1}')"
-
-if [ "${HAS2}" == "kaveh-cluster-ctrl.service" ]; then
-    /bin/systemctl stop kaveh-cluster-ctrl.service || true
-    /bin/systemctl disable kaveh-cluster-ctrl.service || true
-fi
-
-/bin/rm -f /usr/local/bin/kluster-cli || true
-/bin/rm -f /usr/local/bin/kluster-agent || true
-/bin/rm -f /usr/local/lib/kaveh-cluster-ctrl/node/bin/npm
-/bin/rm -f /usr/local/lib/kaveh-cluster-ctrl/node/bin/npx
